requirements.txt
streamlit~=1.39.0
joblib~=1.4.2
numpy~=2.1.2
python-dotenv~=1.0.1
pandas~=2.2.3
matplotlib~=3.9.2
scikit-learn~=1.5.2